move caip stream
